requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
boto3==1.34.140
Pillow==10.4.0
transformers==4.43.3
accelerate==0.33.0
spacy==3.7.6
timm==0.9.16
einops==0.7.0
sentencepiece==0.2.0
safetensors==0.4.3
open-clip-torch==2.24.0
